fixed deleted payment feature and added custom exception classes for it

+++ src/main/java/com/revature/services/PaymentService.java
@@ -2,12 +2,14 @@ package com.revature.services;
 
 import com.revature.daos.AccountDAO;
 import com.revature.daos.PaymentDAO;
+import com.revature.exceptions.payment.PaymentNotFound;
 import com.revature.models.Account;
 import com.revature.models.Payment;
 import com.revature.requests.PaymentDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.swing.text.html.Option;
 import java.util.Optional;
 
 @Service
@@ -22,7 +24,6 @@ public class PaymentService {
     }
 
     /** TODO:
-     *  Delete Payment
      *  Get All By account Id
      *
      * */
@@ -50,12 +51,17 @@ public class PaymentService {
     }
 
     public String deletePayment(Integer accountId, Integer paymentId) throws Exception {
-        String result = "";
-        Account account = accountService.searchById(accountId);
-        paymentDAO.deleteById(paymentId);
-        result = "Account Id : " + account.getAccountId() + " Payment Id : " + paymentId +
-                " Payment Removed";
-        return result;
+            String result = "";
+            Account account = accountService.searchById(accountId);
+            Optional<Payment> checkPayment = paymentDAO.findById(paymentId);
+            if(checkPayment.isEmpty()){
+                throw new PaymentNotFound("PAYMENT ID NOT FOUND");
+            } else {
+                paymentDAO.deleteById(paymentId);
+                result = "Account Id : " + account.getAccountId() + " Payment Id : " + paymentId +
+                        " Payment Removed";
+                return result;
+            }
     }
 }
